♻️ refactor:  WidgetRefExを別ファイルに切り出す
#73
